feat: Move Snowflake to generator mod; Add Generator trait
