+++ README.md
@@ -1,3 +1,5 @@
 # snowflake
 
 A Twitter snowflake algorithm simply implemented in Rust.
+
+[`twitter snowflake `](https://github.com/twitter-archive/snowflake.git)
